Fix handling of `Float` type.

+++ lib/bake/types/float.rb
@@ -20,7 +20,7 @@
 
 module Bake
 	module Types
-		module Integer
+		module Float
 			def self.composite?
 				false
 			end
